Refactor to separate out ldm from async handling
